Converter share.js to typescript
