--- src/lexicon.js
@@ -1,15 +1,0 @@
-/* @flow */
-window.gcexports.globalLexicon = {
-    "val" : { "tk": 1, "name": "VAL", "cls": "function", "length": 2 , "arity": 2 },
-    "key" : { "tk": 1, "name": "KEY", "cls": "function", "length": 2 , "arity": 2 },
-    "len" : { "tk": 1, "name": "LEN", "cls": "function", "length": 1 , "arity": 1 },
-    "concat" : { "tk": 1, "name": "CONCAT", "cls": "function", "length": 1 , "arity": 1 },
-    "add" : { "tk": 1, "name": "ADD", "cls": "function", "length": 2 , "arity": 2 },
-    "mul" : { "tk": 1, "name": "MUL", "cls": "function", "length": 2 , "arity": 2 },
-    "pow" : { "tk": 1, "name": "POW", "cls": "function", "length": 2 , "arity": 2 },
-    "style" : { "tk": 1, "name": "STYLE", "cls": "function", "length": 2, "arity": 2 },
-    "map" : { "tk": 1, "name": "MAP", "cls": "function", "length": 2, "arity": 2 },
-    "apply" : { "tk": 1, "name": "APPLY", "cls": "function", "length": 2, "arity": 2 },
-    "in" : { "tk": 1, "name": "IN", "cls": "function", "length": 0, "arity": 0 },
-    "arg" : { "tk": 1, "name": "ARG", "cls": "function", "length": 1, "arity": 1 }
-}
