+++ src/viewer.js
@@ -1,13 +1,5 @@
 /* Copyright (c) 2017, Art Compiler LLC */
 /* @flow */
-import {
-  assert,
-  message,
-  messages,
-  reserveCodeRange,
-  decodeID,
-  encodeID,
-} from "./share.js";
 import * as React from "react";
 import * as d3 from "d3";
